+++ src/misc/data-generations.ts
@@ -1,6 +1,5 @@
 // https://pokemondb.net/type/old
 // https://pokemondb.net/type
-// TODO: Rename "Generation" to "Game"
 
 export const generations = ["default", "gen2", "gen1"] as const;
 export type Generation = (typeof generations)[number];